requirements.txt
# ── Existing hybrid-search dependencies ──
pandas>=2.0.0
numpy>=1.24.0
rank-bm25>=0.2.2
sentence-transformers>=2.2.2
scikit-learn>=1.3.0

# ── MCP server ──
mcp>=1.0.0

# ── Internet search (no API key needed) ──
ddgs>=0.1.0

# ── Prefab dashboard server ──
flask>=3.0.0

# ── Gemini AI ──
google-genai>=1.0.0
python-dotenv>=1.0.0
